[feat] Add beautified JSON view mode with structured rendering for trace spans
- Replace `rendered-json` mode with `beautified-json` and `decoded-json` modes
- Extract JSON parsing/decoding helpers to `decodedJsonHelpers.ts`
- Move beautified rendering logic to `BeautifiedJsonView.tsx` component
- Add `VIEW_MODES.md` documentation for all display modes
- Update mode selection logic to support `viewModePreset="message"` preference
- Normalize escaped line breaks and unwrap nested stringified JSON in
